+++ examples/statement-playground/index.ts
@@ -1,13 +1,13 @@
-import { createClient } from "@polkadot-api/substrate-client"
 import {
   createStatementSdk,
+  filterPosted,
   getStatementSigner,
   stringToTopic,
+  topicFilter,
   type Statement,
 } from "@polkadot-api/sdk-statement"
-import { getWsProvider } from "polkadot-api/ws"
-import { sign, getPublicKey } from "@scure/sr25519"
 import { Binary, Blake2256 } from "@polkadot-api/substrate-bindings"
+import { getPublicKey, sign } from "@scure/sr25519"
 import { jsonSerialize, mergeUint8 } from "polkadot-api/utils"
 
 // use any key
@@ -16,33 +16,44 @@ const alice = getStatementSigner(getPublicKey(ALICE_SK), "sr25519", (p) =>
   sign(ALICE_SK, p, Blake2256(mergeUint8([ALICE_SK, p]))),
 )
 
-const client = createClient(getWsProvider("ws://127.0.0.1:9936"))
-
-const sdk = createStatementSdk(client.request)
+const sdk = createStatementSdk("ws://127.0.0.1:9936")
 
 // BUILD STATEMENT
 const stmt1: Statement = {
   decryptionKey: stringToTopic("key"),
-  priority: 1,
+  expiry: {
+    timestampSecs: (Date.now() + 3600000) / 1000,
+    sequence: 0,
+  },
   topics: [stringToTopic("1"), stringToTopic("2")],
   data: Binary.fromText("TEST 1"),
 }
 
 // SIGN AND SUBMIT STATEMENT
+const subscription = sdk
+  .getStatement$()
+  .subscribe((r) => console.log("received new statement", r))
+
 const signed1 = await alice.sign(stmt1)
 console.log(await sdk.submit(signed1))
 
 console.log(
+  "posted to 'key' with topics '1' and '2'",
   JSON.stringify(
-    await sdk.getStatements({
-      dest: stringToTopic("key"),
-      topics: [stringToTopic("1"), stringToTopic("2")],
-    }),
+    (
+      await sdk.getStatements(
+        topicFilter([stringToTopic("1"), stringToTopic("2")]),
+      )
+    ).filter(filterPosted(stringToTopic("key"))),
     jsonSerialize,
   ),
 )
 
 // GET ALL STATEMENTS (i.e. `dump`)
-console.log(JSON.stringify(await sdk.getStatements(), jsonSerialize))
+console.log(
+  "all statements",
+  JSON.stringify(await sdk.getStatements(), jsonSerialize),
+)
 
-client.destroy()
+subscription.unsubscribe()
+sdk.destroy()